backend/pdf-extractor-app/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
werkzeug==2.0.3
pandas
PyPDF2
openpyxl
camelot-py
jpype1
tabula-py
pdfplumber
opencv-python
flask-cors
